The first commit of the coYaght system
coYaght is a small 'submarine' drone that contains a Raspberry Pi 3B+, an Arduino, some sensors and some motors to navigate it.
This code creates a HTTP server to be used as the UI.
This project is submitted to the 2019 Junior Achievemt competition by the 1st Junior High School of Gerakas
The project won the Advanced Technology Usage award. The team of the studens worked on the project is:
Zografou Eleni
Antonis Lekkas
Antigoni Bouki
Eleftheria Kalavrou
Startos Amerikanos
Giannis Drosakis
Panagiotis Adamantiades

The supervising teachers are:
Theodoros Kourtesis
Elias Chrysocheris